requirements.txt
# Core ML/DL
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Time series & geospatial
xarray>=2023.0.0
rasterio>=1.3.0
geopandas>=0.14.0

# Explainability
shap>=0.43.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Configuration & utilities
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0

# Optional: API/Dashboard
# streamlit>=1.28.0
# fastapi>=0.104.0
